performance paranoia
This should be slightly faster than a call to strcat(), in theory, since
it's writing up to 512 bytes per iteration from the buffer into the last
significant byte of the array.

My fear is that strcat() makes calls to strlen() internally, making
successive calls to it rather inefficient.

open_memstream() turned up in a quick internet lookup, allowing one to
treat a memory address as a regular file stream, but since it's not even
part of C11, I decided not to use it and instead make calls to
strncpy().

It's possible that just passing an incremented pointer to strcat() would
be clearer, however. I'll look into it sometime, I dunno.