corrected day 11 part 1 in Haskell
The previous one did work for my provided input, but it was an incorrect
solution anyways. This one will step through the provided input and
update values in a map, composing steps as necessary, using a map from
direction to a tuple of possible compose rules.

That map can be reduced as a list comprehension, however, with which I
may or may not replace the current hardcoded stuff.