foldl -> foldr
The distance to the final position doesn't depend on the order in which
the given steps are carried out. The "trail" that the program leves
behind - i.e. the positions it assumes as it makes its way to the final
one - is different, but the end result is the same.

For brevity, I decided to avoid using `flip' or changing the function
types.